FIRMSProject/requirements.txt
pandas
matplotlib
plotly
folium
collection
PyYAML
configparser